--- src/api/data/go.mod
@@ -1,3 +1,0 @@
-module data
-
-go 1.23.2
