--- src/api/services/go.mod
@@ -1,3 +1,0 @@
-module services
-
-go 1.23.2
