+++ src/provisioning/internal/types/users.go
@@ -1,0 +1,8 @@
+package types
+
+type User struct {
+	ID   int
+	Sub  string
+	Name string `json:"name"`
+	Role string `json:"role"`
+}
